+++ lib/util/adobe_strings.py
@@ -68,7 +68,6 @@ DANGEROUS_PDF_KEYS = [
 # Some references are never part of a parent/child relationship in the tree
 NON_TREE_REFERENCES = [
     D,
-    DEST,
     FIRST,
     LAST,
     NEXT,